requirements.txt
# Project dependencies
streamlit>=1.26.0
openai>=1.3.0
requests>=2.31.0
toml>=0.10.2
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
pyautogui>=0.9.54
mss>=10.0.0
streamlit-webrtc>=0.47.1
aiortc>=1.5.0
opencv-python-headless>=4.8.0
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10
